fix problem with storing closed pulls
the closed pulls are now stored and are returning the correct number of pulls. However
the repos with a large number of closed pulls (over 1000) seem to need a second hit of the
url before the correct number of closed pulls is returned. on the first hit it shows the previous
repos value